+++ RadFixes/Helper.cs
@@ -18,5 +18,10 @@ namespace RadFixes
         {
             Traverse.Create(typeof(T)).Field(field).SetValue(value);
         }
+
+        public static object InvokePrivateMethod(this object obj, string method, params object[] parameters)
+        {
+            return AccessTools.Method(obj.GetType(), method).Invoke(obj, parameters);
+        }
     }
 }