+++ RadFixes/Plugin.cs
@@ -1,4 +1,5 @@
 ﻿using BepInEx;
+using BepInEx.Configuration;
 using BepInEx.Logging;
 using HarmonyLib;
 using System.Reflection;
@@ -11,7 +12,10 @@ namespace RadFixes
     {
         public const string PLUGIN_GUID = "com.raddude82.radfixes";
         public const string PLUGIN_NAME = "RadFixes";
-        public const string PLUGIN_VERSION = "1.1.1";
+        public const string PLUGIN_VERSION = "1.1.2";
+
+        internal static ConfigEntry<bool> enableFishingReelFix;
+        internal static ConfigEntry<string> boatCameraMenuZoom;
 
         internal static Plugin instance;
         internal static ManualLogSource logger;
@@ -20,6 +24,9 @@ namespace RadFixes
         {
             instance = this;
             logger = Logger;
+
+            enableFishingReelFix = Config.Bind("Settings", "Enable reel fix", true, "If enabled will rotate the fishing rod reel in the direction it should.");
+            boatCameraMenuZoom = Config.Bind("Settings", "Boat camera settings menu zoom fix", "DisableZoom", new ConfigDescription("While paused in boat camera mode: DisableZoom will disable zooming in and out, DisableMenu will disable the menu, None will leave it as it is.", new AcceptableValueList<string>("DisableZoom", "DisableMenu", "None")));
 
             Harmony.CreateAndPatchAll(Assembly.GetExecutingAssembly(), PLUGIN_GUID);
             SceneManager.sceneLoaded += SceneLoaded;